feat: Add support for displaying history reports on HTML page
Previously, only the information about all reports within a category was displayed on the HTML page. This commit adds support for displaying history reports as well. If there is more than one history report, the date range is shown. Otherwise, only the date of the sole history report is displayed. This enhancement improves the completeness of the information presented on the HTML page.